Function id::account_holder::create_unsigned_credential

pub fn create_unsigned_credential<'a, P: Pairing, C: Curve<Scalar = P::ScalarField>, AttributeType: Attribute<C::Scalar>>(
    context: IpContext<'a, P, C>, 
    id_object: &IdentityObject<P, C, AttributeType>, 
    id_object_use_data: &IdObjectUseData<P, C>, 
    cred_counter: u8, 
    policy: Policy<C, AttributeType>, 
    cred_key_info: CredentialPublicKeys, 
    addr: Option<&AccountAddress>
) -> Result<(UnsignedCredentialDeploymentInfo<P, C, AttributeType>, CommitmentsRandomness<C>)> where
    AttributeType: Clone, 

Generates an unsigned credential deployment info and outputs the randomness used in commitments. The information is meant to be valid in the context of a given identity provider, and global parameter. The ‘cred_counter’ is used to generate a new credential ID. It should be the case that using the output, one can construct an actual credential deployment info, by signing the unsigned challenge.