Function id::account_holder::generate_pio

pub fn generate_pio<P: Pairing, C: Curve<Scalar = P::ScalarField>>(
    context: &IpContext<'_, P, C>, 
    threshold: Threshold, 
    id_use_data: &IdObjectUseData<P, C>, 
    initial_account: &impl InitialAccountDataWithSigning
) -> Option<(PreIdentityObject<P, C>, SigRetrievalRandomness<P>)>

Two flows for creating identities are supported. The first flow involves the creation of an initial account, the second flow does not. The proofs used in the flows are very similar, and therefore factored out in the function generate_pio_common that constructs the common sigma protocol prover used in both flows:

In the version 0 flow, the common prover is AND’ed with a prover showing that RegID = PRF(key_PRF, 0)

In the version 1 flow, the sigma protocol prover is the common prover The generate_pio_common function also produces the bulletproofs rangeproof used in both flows. The version 0 flow is kept for backwards compatibility. Generate a version 0 PreIdentityObject out of the account holder information, the chosen anonymity revoker information, and the necessary contextual information (group generators, shared commitment keys, etc). NB: In this method we assume that all the anonymity revokers in context are to be used.