Function id::anonymity_revoker::reveal_prf_key

pub fn reveal_prf_key<C: Curve>(shares: &[(ArIdentity, Value<C>)]) -> C::Scalar

Reveal the PRF key based on the given shares. It is important to remember that this always succeeds in computing something. It simply does polynomial interpolation. Whether the resulting value is meaningful must be ensured by the caller, e.g., by making sure that the threshold is compatible with the number of shares.