Struct id::id_proof_types::Statement

pub struct Statement<C: Curve, AttributeType: Attribute<C::Scalar>> {
    pub statements: Vec<AtomicStatement<C, AttributeType>>,
}

A statement is a list of atomic statements.

Fields
statements: Vec<AtomicStatement<C, AttributeType>>

The list of atomic statements

Implementations

impl Statement<G1, AttributeKind>

Helper functions for constructing statements

pub fn older_than(self, age: u64) -> Option<Self>

For stating that the user is at least age years old. The functions returns None if

the current year does not fit into a u64, or

the given age is larger than the current year. Otherwise it returns Some(statement) where statement is composed by the statements in self and the “older than” statement.

pub fn younger_than(self, age: u64) -> Option<Self>

For stating that the user is strictly younger than age years old. The functions returns None if

the current year does not fit into a u64, or

the given age is larger than the current year. Otherwise it returns Some(statement) where statement is composed by the statements in self and the “younger than” statement.

pub fn age_in_range(self, lower: u64, upper: u64) -> Option<Self>

For stating that the user’s age in years is in [lower, upper). The functions returns None if

the current year does not fit into a u64, or

the given lower or upper bound is larger than the current year. Otherwise it returns Some(statement) where statement is composed by the statements in self and the “age in range” statement.

pub fn doc_expiry_no_earlier_than(self, lower: AttributeKind) -> Option<Self>

For stating that the user’s document expiry is at least lower. The function returns Some(statement) where statement is composed by the statements in self and the “document expiry no earlier than” statement.

impl<C: Curve, AttributeType: Attribute<C::Scalar>> Statement<C, AttributeType>

pub fn new() -> Self

For constructing the empty statement.

pub fn reveal_attribute(self, attribute_tag: AttributeTag) -> Self

For revealing an attribute. This is resquests the user to reveal the attribute value corresponding to attribute_tag and to prove that the value is the one inside the on-chain commitment. The function returns the statements in self composed with the “reveal attribute” statement.

pub fn in_range(
self,
tag: AttributeTag,
lower: AttributeType,
upper: AttributeType
) -> Self

For stating that an attribute is in [lower, upper). The function returns the statements in self composed with the range statement.

pub fn member_of(self, tag: AttributeTag, set: BTreeSet<AttributeType>) -> Self

For stating that an attribute is in a set. The function returns the statements in self composed with the “member of” statement.

pub fn not_member_of(
self,
tag: AttributeTag,
set: BTreeSet<AttributeType>
) -> Self

For stating that that an attribute does not lie in a set. The function returns the statements in self composed with the “not member of” statement.

pub fn residence_in(self, set: BTreeSet<AttributeType>) -> Option<Self>

For stating that the user’s country of residence is in a set. The function returns Some(statement) where statement is composed by the statements in self and the “residence in” statement.

pub fn residence_not_in(self, set: BTreeSet<AttributeType>) -> Option<Self>

For stating that the user’s country of residence does not lie in set. The function returns Some(statement) where statement is composed by the statements in self and the “residence not in” statement.

pub fn document_issuer_in(self, set: BTreeSet<AttributeType>) -> Option<Self>

For stating that the user’s document issuer is in a set. The function returns Some(statement) where statement is composed by the statements in self and the “document issuer in” statement.

pub fn document_issuer_not_in(
self,
set: BTreeSet<AttributeType>
) -> Option<Self>

For stating that the user’s document issuer does not lie in a set. The function returns Some(statement) where statement is composed by the statements in self and the “document issuer not in” statement.

pub fn nationality_in(self, set: BTreeSet<AttributeType>) -> Option<Self>

For stating that the user’s nationality is in a set. The function returns Some(statement) where statement is composed by the statements in self and the “document issuer in” statement.

pub fn nationality_not_in(self, set: BTreeSet<AttributeType>) -> Option<Self>

For stating that the user’s nationality does not lie in a set. The function returns Some(statement) where statement is composed by the statements in self and the “document issuer not in” statement.

impl<C: Curve, AttributeType: Attribute<C::Scalar>> Statement<C, AttributeType>

pub fn verify(
&self,
challenge: &[u8],
global: &GlobalContext<C>,
credential: &CredId<C>,
commitments: &CredentialDeploymentCommitments<C>,
proofs: &Proof<C, AttributeType>
) -> bool

Function for verifying a proof of a statement. The arguments are

challenge - slice to challenge bytes chosen by the verifier

global - the on-chain cryptographic parameters

credential - the credential for which this statement applies

commitments - the on-chain commitments of the relevant credential The function returns true if the statement is true. If the statement is false, the function returns false with overwhelming probability.