Function id::id_prover::prove_ownership_of_account

pub fn prove_ownership_of_account(
    data: &CredentialData, 
    account: AccountAddress, 
    challenge: &[u8]
) -> AccountOwnershipProof

Function for proving ownership of an account. The parameters are

data - the CredentialData containing the private keys of the prover

account - the account address of the account that the prover claims to own

challenge - a challenge produced by the verifier

The function outputs a proof consisting of signatures on the SHA266 hash of

the account address

0 written as an u64 integer, i.e., eight 0u8’s

the challenge from the verifier

The reason that the 0’s are hashed is to make sure that the hash that is signed cannot coincide with a hash of transaction. When hashing a transaction, the bytestring that is hashed begins with the account address followed by the nonce. For a transaction to be valid, the nonce must be > 0, so by hashing 0 in function below, we are sure that what is signed does not coincide with the hash of a transaction (assuming that SHA256 is collision-resistant).