Function id::id_verifier::verify_account_ownership

pub fn verify_account_ownership(
    public_data: &CredentialPublicKeys, 
    account: AccountAddress, 
    challenge: &[u8], 
    proof: &AccountOwnershipProof
) -> bool

Function for verifying account ownership. The arguments are

public_data - the public keys (and threshold) of the prover. These should be read from chain by looking up the account. If they are not present on chain, the verifier should reject the proof.

account - the account address of the account that the prover claims to own.

challenge - the challenge that the verifier gave to the prover.

proof - the prover’s proof.

The function outputs a bool, indicating whether the proof is correct or not, i.e., it checks the signatures inside the proof.