Module id::identity_provider

Functionality needed by the identity provider. This gathers together the primitives from the rest of the library into a convenient package.

Enums

Reason
Reason for rejecting an identity object request. This is for cryptographic reasons only, real-world identity verification is not handled in this library.

Functions

compute_message

create_initial_cdi
Produce a signature on the initial account data to make a message that is submitted to the chain to create an initial account.

sign_identity_object
Sign the given pre-identity-object to produce an identity object.

validate_request
FIXME: This function does not check that the anonymity revocation parameters make sense. Validate all the proofs in an identity object request.

verify_credentials
Validate the request and sign the identity object.