Function id::identity_provider::create_initial_cdi

pub fn create_initial_cdi<P: Pairing, C: Curve<Scalar = P::ScalarField>, AttributeType: Attribute<C::Scalar>>(
    ip_info: &IpInfo<P>, 
    pub_info_for_ip: PublicInformationForIp<C>, 
    alist: &AttributeList<C::Scalar, AttributeType>, 
    expiry: TransactionTime, 
    ip_cdi_secret_key: &SecretKey
) -> InitialCredentialDeploymentInfo<C, AttributeType>

Produce a signature on the initial account data to make a message that is submitted to the chain to create an initial account.