Function id::identity_provider::validate_id_recovery_request

pub fn validate_id_recovery_request<P: Pairing, C: Curve<Scalar = P::ScalarField>>(
    ip_info: &IpInfo<P>, 
    context: &GlobalContext<C>, 
    request: &IdRecoveryRequest<C>
) -> bool

Verify a ID recovery quest containing proof of knowledge of idCredSec. If the proof verifies, the IDP sends the attribute list and the signature to the prover (the account holder). The argument are

ip_info - Identity provider information containing their IR and verification key that goes in to the protocol context.

context - Global Context containing g such that idCredPub = g^idCredSec. Also goes into the protocol context.

request - the ID recovery containing idCredPub, a timestamp and a proof of knowledge of idCredSec.