Module id::range_proof

Re-export of bulletproofs.

Structs

Generators

Struct containing generators G and H needed for range proofs

RangeProof

Enums

VerificationError

The verifier does two checks. In case verification fails, it can be useful to know which of the checks led to failure.

Functions

prove

This function produces a range proof, i.e. a proof of knowledge of value v_1, v_2, …, v_m that are all in [0, 2^n) that are consistent with commitments V_i to v_i. The arguments are

prove_given_scalars

This function produces a range proof given scalars in a prime field instead of integers. It invokes prove(), documented below.

prove_less_than_or_equal

For proving that a <= b for integers a,b It is assumed that a,b \in [0, 2^n)

verify_efficient

This function verifies a range proof, i.e. a proof of knowledge of value v_1, v_2, …, v_m that are all in [0, 2^n) that are consistent with commitments V_i to v_i. The arguments are

verify_less_than_or_equal

Given commitments to a and b, verify that a <= b It is assumed that b \in [0, 2^n), but it should follow that a \in [0, 2^n) if the proof verifies.