Function id::range_proof::prove_given_scalars

pub fn prove_given_scalars<C, T>(
    transcript: &mut RandomOracle, 
    csprng: &mut T, 
    n: u8, 
    m: u8, 
    v_vec: &[<C as Curve>::Scalar], 
    gens: &Generators<C>, 
    v_keys: &CommitmentKey<C>, 
    randomness: &[Randomness<C>]
) -> Option<RangeProof<C>> where
    C: Curve,
    T: Rng, 

This function produces a range proof given scalars in a prime field instead of integers. It invokes prove(), documented below.

See the documentation of prove below for the meaning of arguments.