Function id::range_proof::verify_efficient

pub fn verify_efficient<C>(
    transcript: &mut RandomOracle, 
    n: u8, 
    commitments: &[Commitment<C>], 
    proof: &RangeProof<C>, 
    gens: &Generators<C>, 
    v_keys: &CommitmentKey<C>
) -> Result<(), VerificationError> where
    C: Curve, 

This function verifies an aggregated range proof, i.e., a proof of knowledge of values v_1, v_2, ..., v_m in [0, 2^n) that are consistent with commitments V_i to v_i. The arguments are

n - the number n such that each v_i is claimed to be in [0, 2^n) by the prover

commitments - commitments V_i to each v_i

proof - the range proof

gens - generators containing vectors G and H both of length at least nm (bold g,h in bluepaper)

v_keys - commitment keys B and B_tilde (g,h in bluepaper)

Note: The bulletproof paper also describes an optimized verification method that integrates the exponentiations from the inner-product verification into the range proof verification using the Schwartz–Zippel lemma. We had implemented this and compared the performance, but since the performance gains were negligible and modularity much worse, we do not use this here.