Function id::secret_sharing::share

pub fn share<C: Curve, P: Into<u64>, I: IntoIterator<Item = P> + ExactSizeIterator, R: Rng>(
    secret: &C::Scalar, 
    points: I, 
    revealing_threshold: Threshold, 
    csprng: &mut R
) -> SharingData<C>

Revealing Threshold must be at least 1. We assume the points are anything that can be converted to a u64. They should be distinct, and the polynomial will be evaluated at points p in points. The preconditions are not checked.