Struct id::sigma_protocols::common::AndAdapter

pub struct AndAdapter<P1, P2> {
    pub first: P1,
    pub second: P2,
}

An adapter to combine multiple provers or multiple verifiers. The marker type C is for convenience in use with the SigmaProtocolProver/Verifier traits below.

Fields
first: P1second: P2
Implementations

impl<P1: SigmaProtocol, P2: SigmaProtocol> AndAdapter<P1, P2>

pub fn add_prover<P3: SigmaProtocol>(
self,
additional_prover: P3
) -> AndAdapter<AndAdapter<P1, P2>, P3>

Extend the current adapter with a new prover.