Trait id::sigma_protocols::common::SigmaProtocol

pub trait SigmaProtocol: Sized {
    type CommitMessage: Serial;
    type ProverWitness: Serialize;
    type ProtocolChallenge;
    type ProverState;
    type SecretData;
    fn public(&self, ro: &mut RandomOracle);

    fn get_challenge(&self, challenge: &Challenge) -> Self::ProtocolChallenge;

    fn commit_point<R: Rng>(
        &self, 
        csprng: &mut R
    ) -> Option<(Self::CommitMessage, Self::ProverState)>;

    fn generate_witness(
        &self, 
        secret: Self::SecretData, 
        state: Self::ProverState, 
        challenge: &Self::ProtocolChallenge
    ) -> Option<Self::ProverWitness>;

    fn extract_point(
        &self, 
        challenge: &Self::ProtocolChallenge, 
        witness: &Self::ProverWitness
    ) -> Option<Self::CommitMessage>;
}

The common data known to the prover and verifier, i.e., public values and coefficients.

Associated Types

type CommitMessage: Serial

Prover’s commit message.

type ProverWitness: Serialize

Prover’s witness.

type ProtocolChallenge

Inherent challenge type. This needs to be convertible from the Challenge type above.

type ProverState

Prover’s state after the first message. Essentially any randomness they’ve chosen to generate the commit message.

type SecretData

The prover’s secret data.

Required methods

fn public(&self, ro: &mut RandomOracle)

Given a random oracle, feed it all of the public input of this instance.

fn get_challenge(&self, challenge: &Challenge) -> Self::ProtocolChallenge

fn commit_point<R: Rng>(
&self,
csprng: &mut R
) -> Option<(Self::CommitMessage, Self::ProverState)>

First message generated by the prover. We allow this function to return ‘None’ if the inputs are inconsistent. The arguments are

self – the prover’s public and secret data

secret – secret data known to the prover only

csprng – a cryptographically secure random number generator

fn generate_witness(
&self,
secret: Self::SecretData,
state: Self::ProverState,
challenge: &Self::ProtocolChallenge
) -> Option<Self::ProverWitness>

Third message generated by the prover. We allow this function to return ‘None’ if the inputs are inconsistent.

self – the prover’s public and secret data

secret – secret data known to the prover only

state – the state returned from the ‘commit_point’ call

challenge – computed challenge This function is pure and deterministic.

fn extract_point(
&self,
challenge: &Self::ProtocolChallenge,
witness: &Self::ProverWitness
) -> Option<Self::CommitMessage>

Combine the public outputs from the third message with the challenge and produce a commitment for verification.

challenge – computed challenge

self – the verifier’s public data This function is pure and deterministic. It is allowed to return ‘None’ if some of the input data is malformed, e.g., vectors of inconsistent lengths.