Struct id::types::PreIdentityProof

pub struct PreIdentityProof<P: Pairing, C: Curve<Scalar = P::ScalarField>> {
    pub challenge: Challenge,
    pub id_cred_sec_witness: Witness<C>,
    pub commitments_same_proof: Witness<C>,
    pub commitments_prf_same: Witness<P::G1, C>,
    pub prf_regid_proof: Witness<C>,
    pub proof_acc_sk: AccountOwnershipProof,
    pub bulletproofs: Vec<RangeProof<C>>,
}

Proof that the data sent to the identity provider is well-formed.

Fields
challenge: Challenge

Challenge for the combined proof. This includes the three proofs below, and additionally also the proofs in IpArData.

id_cred_sec_witness: Witness<C>

Witness to the proof of konwledge of IdCredSec.

commitments_same_proof: Witness<C>

Witness to the proof that cmm_sc and id_cred_pub are hiding the same id_cred_sec.

commitments_prf_same: Witness<P::G1, C>

Witness to the proof that cmm_prf and the second commitment to the prf key (hidden in cmm_prf_sharing_coeff) are hiding the same value.

prf_regid_proof: Witness<C>

Witness to the proof that reg_id = PRF(prf_key, 0)

proof_acc_sk: AccountOwnershipProof

Signature on the public information for the IP from the account holder

bulletproofs: Vec<RangeProof<C>>

Bulletproofs for showing that chunks are small so that encryption of the prf key can be decrypted