Trait id::types::Attribute

pub trait Attribute<F: Field>: Clone + Sized + Send + Sync + Display + Serialize {
    fn to_field_element(&self) -> F;
}

An abstraction of an attribute. In the id library internals the only thing we care about attributes is that they can be encoded as field elements. The meaning of attributes is then assigned at the outer layers when the library is used. In order to make the library as generic (and ultimately simple) as possible this trait is used.

Required methods

fn to_field_element(&self) -> F

Convert an attribute to a field element

Loading content...