Trait id::types::InitialAccountDataWithSigning

pub trait InitialAccountDataWithSigning: PublicInitialAccountData {
    fn sign_public_information_for_ip<C: Curve>(
        &self, 
        info: &PublicInformationForIp<C>
    ) -> BTreeMap<KeyIndex, AccountOwnershipSignature>;
}

A helper trait to allow signing PublicInformationForIP in an implementation that does not give access to the secret keys.

Required methods

fn sign_public_information_for_ip<C: Curve>(
&self,
info: &PublicInformationForIp<C>
) -> BTreeMap<KeyIndex, AccountOwnershipSignature>

Sign a PublicInformationForIP structure with the secret keys that matches the public keys, which the structure provides. NB: the function should, for each secret key, sign the sha256 hash of the structure’s serialization.

Loading content...