Trait id::types::PublicCredentialData

pub trait PublicCredentialData {
    fn get_public_keys(&self) -> BTreeMap<KeyIndex, VerifyKey>;

    fn get_threshold(&self) -> SignatureThreshold;

    fn get_cred_key_info(&self) -> CredentialPublicKeys { ... }
}

A helper trait to access the public parts of the CredentialData structure. We use this to allow implementations that does not give or have access to the secret keys. NB: the threshold should be at most the number of keypairs.

Required methods

fn get_public_keys(&self) -> BTreeMap<KeyIndex, VerifyKey>

Get the public keys of the credential

fn get_threshold(&self) -> SignatureThreshold

Get the signature threshold of the account.

Loading content...
Provided methods

fn get_cred_key_info(&self) -> CredentialPublicKeys

Get the CredentialPublicKeys struct directly

Loading content...