Function id::utils::credential_hash_to_sign

pub fn credential_hash_to_sign<P: Pairing, C: Curve<Scalar = P::ScalarField>, AttributeType: Attribute<C::Scalar>>(
    values: &CredentialDeploymentValues<C, AttributeType>, 
    proofs: &IdOwnershipProofs<P, C>, 
    new_or_existing: &Either<TransactionTime, AccountAddress>
) -> Vec<u8>

Compute the hash of the credential deployment that should be signed by the account keys for deployment. If new_or_existing is Left then this credential will create a new account, and the transaction expiry should be signed. otherwise it is going to be deployed to the given account address, and the address should be signed.