Crate id

Re-exports

pub use pedersen_scheme as pedersen_commitment;

pub use curve_arithmetic;

pub use elgamal;

pub use dodis_yampolskiy_prf;

Modules

account_holder

Functionality needed by the account holder, either when interacting with the identity provider, or when interacting with the chain.

anonymity_revoker

Functionality needed by the anonymity revoker.

chain

Functionality needed by the chain to verify credential deployments.

constants

Collection of constants that fix choices, be it values or types, that are used in various places.

id_prover

id_verifier

identity_provider

Functionality needed by the identity provider. This gathers together the primitives from the rest of the library into a convenient package.

range_proof

Re-export of bulletproofs.

secret_sharing

Implementation of Shamir secret sharing.

sigma_protocols

Implementations of sigma protocols needed by the identity layer of Concordium.

types

Main types used in the identity layer, and their serialization implementations.

utils

A collection of auxiliary functions that don’t belong anywhere else.