Module id::identity_provider

Functionality needed by the identity provider. This gathers together the primitives from the rest of the library into a convenient package.

Enums

Reason

Reason for rejecting an identity object request. This is for cryptographic reasons only, real-world identity verification is not handled in this library.

Functions

compute_message

create_initial_cdi

Produce a signature on the initial account data to make a message that is submitted to the chain to create an initial account.

sign_identity_object

Sign the given pre-identity-object to produce a version 0 identity object. The inputs are

sign_identity_object_v1

Sign the given pre-identity-object to produce a version 1 identity object. The inputs are

validate_id_recovery_request

Verify a ID recovery quest containing proof of knowledge of idCredSec. If the proof verifies, the IDP sends the attribute list and the signature to the prover (the account holder). The argument are

validate_request

The validation of the two versions of identity object requests are very similar, and therefore the common validation parts of the two flows are factored out in the function validate_request_common. It produces the common sigma protocol verifier and witnesses that are used in both validate_request and validate_request_v1:

validate_request_v1

Validate all the proofs in a version 1 identity object request. This is for the flow, where no initial account creation is involved.

verify_credentials

Validate the request and sign the version 0 identity object.

verify_credentials_v1

Validate the request and sign the version 1 identity object.