Trait id::types::CredentialDataWithSigning

pub trait CredentialDataWithSigning: PublicCredentialData {
    fn sign<P: Pairing, C: Curve<Scalar = P::ScalarField>, AttributeType: Attribute<C::Scalar>>(
        &self, 
        message_expiry: &Either<TransactionTime, AccountAddress>, 
        unsigned_cred_info: &UnsignedCredentialDeploymentInfo<P, C, AttributeType>
    ) -> BTreeMap<KeyIndex, AccountOwnershipSignature>;
}

A helper trait to allow signing PublicInformationForIP in an implementation that does not give access to the secret keys.

Required methods

fn sign<P: Pairing, C: Curve<Scalar = P::ScalarField>, AttributeType: Attribute<C::Scalar>>(
&self,
message_expiry: &Either<TransactionTime, AccountAddress>,
unsigned_cred_info: &UnsignedCredentialDeploymentInfo<P, C, AttributeType>
) -> BTreeMap<KeyIndex, AccountOwnershipSignature>

Sign the credential deployment information, proving ownership of account keys. If the first argument is Right(addr) then this credential is intended for an existing account. If the first argument is Left(tt) then the credential is going to create a new account, and the payload is the expiry time of the transaction.

Loading content...